feat(shared): VendaDerivadaCard extraído pra @/Components/shared (OficinaAuto integra · ADR 0192)
Onda 7 da integração Vendas × Oficina A1 KB-9.75 (ADR 0192) — leva o card
"Esta OS gerou venda #V-NNNN" do drawer Repair/ProducaoOficina pro drawer
ServiceOrderSheet do módulo OficinaAuto, via shared component cross-módulo.

Backend (PR #1530) já criava Transaction derivada via ServiceOrderObserver
quando OS OficinaAuto transitiona pra status='concluida' (extensão do
JobSheetObserver mãe ADR 0192). Faltava a UI no Sheet OficinaAuto.

Decisão arquitetural Wagner 2026-05-25: extrair pra shared (DRY) em vez de
duplicar componente in-file. Wave futura (FASE C) toca 1 lugar só.

## Arquivos

**Shared component novo:**
- `resources/js/Components/shared/VendaDerivadaCard.tsx` (+349 LOC) — componente
  + 4 interfaces (VendaItem, VendaItemsSummary, VendaFiscal, VendaDerivada)
  exportadas. Preserva 100% comportamento FASE B do PR #1516 (breakdown peças/
  serviço, badge fiscal NF-e 4 estados, items list expandível, Web Share API).

**Repair (refactor mecânico — comportamento preservado):**
- `resources/js/Pages/Repair/ProducaoOficina/Index.tsx` (-302 LOC net)
  — import shared + remove interfaces/componente inline.
- `Modules/Repair/Tests/Feature/ProducaoOficinaOnda5CompartilharTest.php` +
  `ProducaoOficinaFaseBVendaDerivadaCardTest.php` — guards estruturais agora
  apontam pro shared (via __DIR__ pattern em vez de base_path).

**OficinaAuto (nova integração):**
- `Modules/OficinaAuto/Http/Controllers/ServiceOrderController.php` — eager-loads
  `transaction:id,business_id,invoice_no,final_total,transaction_date` + entrega
  `venda_derivada` no JSON `wantsJson()` via helper `shapeVendaDerivada()` privado
  (V0 core 4 campos · FASE B items/fiscal fica pra wave futura).
- `resources/js/Pages/OficinaAuto/ServiceOrders/_components/ServiceOrderSheet.tsx`
  — interface ServiceOrderDetail ganha `venda_derivada?: VendaDerivada | null`,
  import shared, renderiza card no topo do drawer body.
- `Modules/OficinaAuto/Tests/Feature/ServiceOrderSheetVendaDerivadaGuardTest.php`
  (3 guards estruturais · rodam sempre, sem DB) + `ServiceOrderSheetVendaDerivadaTest.php`
  (3 feature tests · MySQL only — null/populated/multi-tenant biz=1 vs biz=2).

## Validação local (worktree)

```
✓ Onda 5 share guards: 9 passed
✓ FASE B card guards: 15 passed
✓ OficinaAuto VendaDerivadaCard guards: 3 passed
- OficinaAuto VendaDerivadaCard feature: 3 skipped (SQLite — esperado, CI MySQL)
```

## Multi-tenant Tier 0 (ADR 0093 IRREVOGÁVEL)

Transaction load via belongsTo respeita global scope herdado. Test
`Multi-tenant Tier 0: OS biz=1 com transaction biz=2 retorna null` cobre o
cenário adversarial (FK forçada cross-tenant em DB level).

## Não-goals desta wave (futura)

- FASE B no OficinaAuto: items_list + items_summary + fiscal NF-e exigem
  join sell_lines + NfeBrasil. Backend Repair já tem (`buildVendaDerivadaPayload`).
  Quando trouxer pra OficinaAuto, extrair `App\Services\VendaDerivadaPayloadService`
  pra DRY com Repair.
- Backfill de OS já concluídas pré-PR #1530 (sem transaction_id) — Wagner pede
  via artisan command quando Martinho reportar.

Refs:
- ADR 0192 (auto-faturar OS→Venda Observer)
- ADR 0093 (multi-tenant Tier 0)
- ADR 0137 (OficinaAuto qualificada)
- PR #1504 (Onda 5 card original Repair)
- PR #1516 (FASE B evolution items/fiscal)
- PR #1530 (ServiceOrderObserver OficinaAuto · mergeado hoje)
- prototipo-ui/oficina-page.jsx linhas 392-458 (fonte Cowork)
- memory/sessions/2026-05-25-como-integrar-oficina-prototipo-3-itens.md
