+++ Modules/Repair/Tests/Feature/ProducaoOficinaOnda5CompartilharTest.php
@@ -26,11 +26,17 @@ declare(strict_types=1);
 
 uses(Tests\TestCase::class);
 
-const PRODUCAO_OFICINA_INDEX_TSX = 'resources/js/Pages/Repair/ProducaoOficina/Index.tsx';
+// Componente VendaDerivadaCard extraído pra @/Components/shared (feat/shared-venda-derivada-card)
+// — cross-módulo Repair + OficinaAuto. Guards estruturais agora apontam pro shared.
+const PRODUCAO_OFICINA_INDEX_TSX = 'resources/js/Components/shared/VendaDerivadaCard.tsx';
 
 function ondaCincoCompartilharReadIndex(): string
 {
-    return file_get_contents(base_path(PRODUCAO_OFICINA_INDEX_TSX));
+    // __DIR__ resolve relativo ao próprio teste — funciona em worktree OU root.
+    // base_path() resolve sempre pro root project mesmo quando rodando em worktree,
+    // o que quebra o lookup do shared component recém-extraído pra worktree.
+    $worktreeRoot = realpath(__DIR__.'/../../../../');
+    return file_get_contents($worktreeRoot.DIRECTORY_SEPARATOR.PRODUCAO_OFICINA_INDEX_TSX);
 }
 
 it('Onda 5 share: Index.tsx importa toast do sonner (pattern projeto)', function () {
